LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/window-open-self-during-load-expected.txt



FAIL No replace during load, triggered by window.open(_self) on an iframe assert_equals: history.length must increase expected 14 but got 13
